Move fn_to_numeric_cast_with_truncation to its own module
